finish chain_rand implementation
