+++ blockchain/chain/src/store/chain_store.rs
@@ -8,15 +8,12 @@ use async_std::channel::{self, bounded, Receiver};
 use async_std::sync::RwLock;
 use async_std::task;
 use beacon::{BeaconEntry, IGNORE_DRAND_VAR};
-use blake2b_simd::Params;
 use blocks::{Block, BlockHeader, FullTipset, Tipset, TipsetKeys, TxMeta};
-use byteorder::{BigEndian, WriteBytesExt};
 use cid::Cid;
 use cid::Code::Blake2b256;
 use clock::ChainEpoch;
 use crossbeam::atomic::AtomicCell;
-use crypto::DomainSeparationTag;
-use encoding::{blake2b_256, de::DeserializeOwned, from_slice, Cbor};
+use encoding::{de::DeserializeOwned, from_slice, Cbor};
 use forest_car::CarHeader;
 use forest_ipld::recurse_links;
 use futures::AsyncWrite;
@@ -32,7 +29,6 @@ use num_traits::Zero;
 use serde::Serialize;
 use state_tree::StateTree;
 use std::error::Error as StdError;
-use std::io::Write;
 use std::sync::Arc;
 use std::{
     collections::{HashMap, HashSet, VecDeque},
@@ -290,153 +286,153 @@ where
         }
     }
 
-    /// Gets 32 bytes of randomness for ChainRand paramaterized by the DomainSeparationTag, ChainEpoch,
-    /// Entropy from the ticket chain.
-    pub async fn get_chain_randomness(
-        &self,
-        blocks: &TipsetKeys,
-        pers: DomainSeparationTag,
-        round: ChainEpoch,
-        entropy: &[u8],
-        lookback: bool,
-    ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
-        let ts = self.tipset_from_keys(blocks).await?;
+    // /// Gets 32 bytes of randomness for ChainRand paramaterized by the DomainSeparationTag, ChainEpoch,
+    // /// Entropy from the ticket chain.
+    // pub async fn get_chain_randomness(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     pers: DomainSeparationTag,
+    //     round: ChainEpoch,
+    //     entropy: &[u8],
+    //     lookback: bool,
+    // ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
+    //     let ts = self.tipset_from_keys(blocks).await?;
 
-        if round > ts.epoch() {
-            return Err("cannot draw randomness from the future".into());
-        }
+    //     if round > ts.epoch() {
+    //         return Err("cannot draw randomness from the future".into());
+    //     }
 
-        let search_height = if round < 0 { 0 } else { round };
+    //     let search_height = if round < 0 { 0 } else { round };
 
-        let rand_ts = self.tipset_by_height(search_height, ts, lookback).await?;
+    //     let rand_ts = self.tipset_by_height(search_height, ts, lookback).await?;
 
-        draw_randomness(
-            rand_ts
-                .min_ticket()
-                .ok_or("No ticket exists for block")?
-                .vrfproof
-                .as_bytes(),
-            pers,
-            round,
-            entropy,
-        )
-    }
+    //     draw_randomness(
+    //         rand_ts
+    //             .min_ticket()
+    //             .ok_or("No ticket exists for block")?
+    //             .vrfproof
+    //             .as_bytes(),
+    //         pers,
+    //         round,
+    //         entropy,
+    //     )
+    // }
 
-    /// network v0-12
-    pub async fn get_chain_randomness_v1(
-        &self,
-        blocks: &TipsetKeys,
-        pers: DomainSeparationTag,
-        round: ChainEpoch,
-        entropy: &[u8],
-    ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
-        self.get_chain_randomness(blocks, pers, round, entropy, true)
-            .await
-    }
+    // /// network v0-12
+    // pub async fn get_chain_randomness_v1(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     pers: DomainSeparationTag,
+    //     round: ChainEpoch,
+    //     entropy: &[u8],
+    // ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
+    //     self.get_chain_randomness(blocks, pers, round, entropy, true)
+    //         .await
+    // }
 
-    /// network v13 onwards
-    pub async fn get_chain_randomness_v2(
-        &self,
-        blocks: &TipsetKeys,
-        pers: DomainSeparationTag,
-        round: ChainEpoch,
-        entropy: &[u8],
-    ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
-        self.get_chain_randomness(blocks, pers, round, entropy, false)
-            .await
-    }
+    // /// network v13 onwards
+    // pub async fn get_chain_randomness_v2(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     pers: DomainSeparationTag,
+    //     round: ChainEpoch,
+    //     entropy: &[u8],
+    // ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
+    //     self.get_chain_randomness(blocks, pers, round, entropy, false)
+    //         .await
+    // }
 
-    /// network v0-12; with lookback
-    pub async fn get_beacon_randomness_v1(
-        &self,
-        blocks: &TipsetKeys,
-        pers: DomainSeparationTag,
-        round: ChainEpoch,
-        entropy: &[u8],
-    ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
-        self.get_beacon_randomness(blocks, pers, round, entropy, true)
-            .await
-    }
+    // /// network v0-12; with lookback
+    // pub async fn get_beacon_randomness_v1(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     pers: DomainSeparationTag,
+    //     round: ChainEpoch,
+    //     entropy: &[u8],
+    // ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
+    //     self.get_beacon_randomness(blocks, pers, round, entropy, true)
+    //         .await
+    // }
 
-    /// network v13; without lookback
-    pub async fn get_beacon_randomness_v2(
-        &self,
-        blocks: &TipsetKeys,
-        pers: DomainSeparationTag,
-        round: ChainEpoch,
-        entropy: &[u8],
-    ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
-        self.get_beacon_randomness(blocks, pers, round, entropy, false)
-            .await
-    }
+    // /// network v13; without lookback
+    // pub async fn get_beacon_randomness_v2(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     pers: DomainSeparationTag,
+    //     round: ChainEpoch,
+    //     entropy: &[u8],
+    // ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
+    //     self.get_beacon_randomness(blocks, pers, round, entropy, false)
+    //         .await
+    // }
 
-    /// network v14 onwards
-    pub async fn get_beacon_randomness_v3(
-        &self,
-        blocks: &TipsetKeys,
-        pers: DomainSeparationTag,
-        round: ChainEpoch,
-        entropy: &[u8],
-    ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
-        if round < 0 {
-            return self
-                .get_beacon_randomness_v2(blocks, pers, round, entropy)
-                .await;
-        }
+    // /// network v14 onwards
+    // pub async fn get_beacon_randomness_v3(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     pers: DomainSeparationTag,
+    //     round: ChainEpoch,
+    //     entropy: &[u8],
+    // ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
+    //     if round < 0 {
+    //         return self
+    //             .get_beacon_randomness_v2(blocks, pers, round, entropy)
+    //             .await;
+    //     }
 
-        let beacon_entry = self.extract_beacon_entry_for_epoch(blocks, round).await?;
-        draw_randomness(beacon_entry.data(), pers, round, entropy)
-    }
+    //     let beacon_entry = self.extract_beacon_entry_for_epoch(blocks, round).await?;
+    //     draw_randomness(beacon_entry.data(), pers, round, entropy)
+    // }
 
-    /// Gets 32 bytes of randomness for ChainRand paramaterized by the DomainSeparationTag, ChainEpoch,
-    /// Entropy from the latest beacon entry.
-    pub async fn get_beacon_randomness(
-        &self,
-        blocks: &TipsetKeys,
-        pers: DomainSeparationTag,
-        round: ChainEpoch,
-        entropy: &[u8],
-        lookback: bool,
-    ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
-        let rand_ts = self
-            .get_beacon_randomness_tipset(blocks, round, lookback)
-            .await?;
-        let be = self.latest_beacon_entry(&rand_ts).await?;
-        draw_randomness(be.data(), pers, round, entropy)
-    }
+    // /// Gets 32 bytes of randomness for ChainRand paramaterized by the DomainSeparationTag, ChainEpoch,
+    // /// Entropy from the latest beacon entry.
+    // pub async fn get_beacon_randomness(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     pers: DomainSeparationTag,
+    //     round: ChainEpoch,
+    //     entropy: &[u8],
+    //     lookback: bool,
+    // ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
+    //     let rand_ts = self
+    //         .get_beacon_randomness_tipset(blocks, round, lookback)
+    //         .await?;
+    //     let be = self.latest_beacon_entry(&rand_ts).await?;
+    //     draw_randomness(be.data(), pers, round, entropy)
+    // }
 
-    pub async fn extract_beacon_entry_for_epoch(
-        &self,
-        blocks: &TipsetKeys,
-        epoch: ChainEpoch,
-    ) -> Result<BeaconEntry, Box<dyn std::error::Error>> {
-        let rand_ts = self
-            .get_beacon_randomness_tipset(blocks, epoch, false)
-            .await?;
-        // let round = self.beacon.beacon_for_epoch(epoch);
-        Err("unimplemented".into())
-    }
+    // pub async fn extract_beacon_entry_for_epoch(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     epoch: ChainEpoch,
+    // ) -> Result<BeaconEntry, Box<dyn std::error::Error>> {
+    //     let rand_ts = self
+    //         .get_beacon_randomness_tipset(blocks, epoch, false)
+    //         .await?;
+    //     // let round = self.beacon.beacon_for_epoch(epoch);
+    //     Err("unimplemented".into())
+    // }
 
-    pub async fn get_beacon_randomness_tipset(
-        &self,
-        blocks: &TipsetKeys,
-        round: ChainEpoch,
-        lookback: bool,
-    ) -> Result<Arc<Tipset>, Box<dyn std::error::Error>> {
-        let ts = self.tipset_from_keys(blocks).await?;
+    // pub async fn get_beacon_randomness_tipset(
+    //     &self,
+    //     blocks: &TipsetKeys,
+    //     round: ChainEpoch,
+    //     lookback: bool,
+    // ) -> Result<Arc<Tipset>, Box<dyn std::error::Error>> {
+    //     let ts = self.tipset_from_keys(blocks).await?;
 
-        if round > ts.epoch() {
-            return Err("cannot draw randomness from the future".into());
-        }
+    //     if round > ts.epoch() {
+    //         return Err("cannot draw randomness from the future".into());
+    //     }
 
-        let search_height = if round < 0 { 0 } else { round };
+    //     let search_height = if round < 0 { 0 } else { round };
 
-        let res = self.tipset_by_height(search_height, ts, lookback).await;
-        match res {
-            Ok(rand_ts) => return Ok(rand_ts),
-            Err(e) => return Err(e.into()),
-        };
-    }
+    //     let res = self.tipset_by_height(search_height, ts, lookback).await;
+    //     match res {
+    //         Ok(rand_ts) => return Ok(rand_ts),
+    //         Err(e) => return Err(e.into()),
+    //     };
+    // }
 
     /// Finds the latest beacon entry given a tipset up to 20 tipsets behind
     pub async fn latest_beacon_entry(&self, ts: &Tipset) -> Result<BeaconEntry, Error> {
@@ -890,23 +886,23 @@ where
     Ok(())
 }
 
-/// Computes a pseudorandom 32 byte Vec.
-pub fn draw_randomness(
-    rbase: &[u8],
-    pers: DomainSeparationTag,
-    round: ChainEpoch,
-    entropy: &[u8],
-) -> Result<[u8; 32], Box<dyn std::error::Error>> {
-    let mut state = Params::new().hash_length(32).to_state();
-    state.write_i64::<BigEndian>(pers as i64)?;
-    let vrf_digest = blake2b_256(rbase);
-    state.write_all(&vrf_digest)?;
-    state.write_i64::<BigEndian>(round as i64)?;
-    state.write_all(entropy)?;
-    let mut ret = [0u8; 32];
-    ret.clone_from_slice(state.finalize().as_bytes());
-    Ok(ret)
-}
+// /// Computes a pseudorandom 32 byte Vec.
+// pub fn draw_randomness(
+//     rbase: &[u8],
+//     pers: DomainSeparationTag,
+//     round: ChainEpoch,
+//     entropy: &[u8],
+// ) -> Result<[u8; 32], Box<dyn std::error::Error>> {
+//     let mut state = Params::new().hash_length(32).to_state();
+//     state.write_i64::<BigEndian>(pers as i64)?;
+//     let vrf_digest = blake2b_256(rbase);
+//     state.write_all(&vrf_digest)?;
+//     state.write_i64::<BigEndian>(round as i64)?;
+//     state.write_all(entropy)?;
+//     let mut ret = [0u8; 32];
+//     ret.clone_from_slice(state.finalize().as_bytes());
+//     Ok(ret)
+// }
 
 /// Returns a vector of cids from provided root cid
 fn read_amt_cids<DB>(db: &DB, root: &Cid) -> Result<Vec<Cid>, Error>
