+++ node/rpc/src/chain_api.rs
@@ -271,7 +271,6 @@ where
     let entropy = entropy.unwrap_or_default();
     Ok(data
         .state_manager
-        .chain_store()
         .get_chain_randomness(
             &tsk,
             DomainSeparationTag::from_i64(pers).ok_or("invalid DomainSeparationTag")?,
@@ -295,7 +294,6 @@ where
 
     Ok(data
         .state_manager
-        .chain_store()
         .get_beacon_randomness(
             &tsk,
             DomainSeparationTag::from_i64(pers).ok_or("invalid DomainSeparationTag")?,
